- Added identity functions #765
